Add a parser for UnicodeData.txt entries
I have also added the UnicodeData.txt itself.  The overall policy
going forward is likely going to be to add the files used by the
program, and nothing else.  This keeps the repository as lean as
possible, and self-contained.

Naming is unfortunate throughout; I hope to fix that later as I add a
proper structure to the application.