Add a simple trie generator
Of note, I have only confirmed that the results look roughly correct
and nothing bottoms out; there are no tests in place, so no guarantees
that the code is actually correct.

I intend to add a more detailed description in comments at a later
time, but for the time being, the central idea is to

1. Partition a code point (which is a 21-bit number) into bit ranges.
Since there are *17* planes, and actually a code point is a number in
a range 0..0x110000, most senior 5 bits are sort of stuck together,
and largest possible partitioning is actually at 16 bits (i.e. plane
level).

2. Represent each Unicode property table as a trie over said
partitioning.  The partitioning can, in principle, be different for
each property.

3. Deduplicate identical sub-tries to save space.

The mkTrie function in this commit does roughly this.  Full version
will have to accommodate generation of multiple tries and choosing the
most compact one, but this is the prototype.