Switch to Enum for conversions to and from Int
I don't know if it is any faster than fromIntegral for release builds,
but probably makes a difference in debug.  The central issue here is
that *Enum functions fix one type to Int, and dispatches on the other,
whereas fromIntegral is generic in both, and, in general, uses Integer
as an intermediary.  There are rewrite rules in place to specialise it
for specific types, but I'm not sure how comprehensive they are, and,
anyway, they don't fire in debug builds.