Use narrower types for tables
There are several components in this commit, smashed together because
they happen work for the same purpose.

1. The annotation system for Trie, borrowed from Trees that Grow[1].
We don't need to add alternatives, so that's not here, but other than
that this is basically a faithful implementation of the approach
outlined in the paper.

2. The actual type system.  We're only dealing with integral types so
far, so the only constraint is that the type is big enough to fit all
the values requested.  Of those, I weakly prefer signed integers,
because we're eventually converting everything to signed integers, so
that might be more straightforward (esp. since GHC represents all
signed integers identically).

I stop at 32 bits and throw error, because everything ends up in Int
eventually, which is 32 bit on 32 bit systems.  Note that the system
used to generate code and the one used to run it could be drastically
different, so we can't just use (maxBound :: Int) or something to
differentiate.  We should never need more than 21 bits anyway.

3. The warning suppression.  An alternative would be to use qualified
imports, but meh.  This is all autogenerated anyway.

4. Const.  The fact that the tables weren't const from the start was
an oversight that I just happened to fix since I was touching that
code anyway.

[1]: Trees that grow.  Shayan Najd, Simon Peyton Jones.
Journal of Universal Computer Science (JUCS), January 2017,
Vol 23: pp. 47-62
doi: 10.3217/jucs-023-01-0042
https://www.microsoft.com/en-us/research/publication/trees-that-grow/