Add benchmark
Based on the text of Universal Declaration of Human Rights, courtesy
of UDHR in Unicode project[1].

The text, while long, is supposed to provide a sample text in most
scripts and languages supported by Unicode (indeed, that is its
explicit purpose), and thus is a natural fit for what we're trying to
do here.

The best news from all of that is that UCD is 3x faster than
Data.Char.

[1]: https://www.unicode.org/udhr/aggregates.html