Make TrieDesc a bifunctor
This required a slight rearrangement of the type parameters.
Technically, TrieDesc is a trifunctor, but there's no class for that.
It is also bifoldable and bitraversable, but those are less
immediately useful.

On a motivation side, this is preparation for separating
type-assignment of the property values proper from type-assignment of
table indices (b/c the former can be shared between different trie
configurations), for which I need direct access to bottom annotation.