Add ICU-based test
For the time being, test for Name excludes algorithmically generated
ranges, as these are not yet supported.

I happen to have latest version of ICU available in nixpkgs, which is
why these are rather straightforward; I do not yet have a good plan
for dealing with older ICU versions that do not support the latest
version of Unicode.