Add proper hangul syllables handling
This completes the handling of Name property.  Since the range of
values in which hangul jamo live is so small, I opted to use a simple
array instead of a usual trie; the whole contraption fits int 500
bytes, which is fairly small by itself and vanishingly small compared
to the full names table.