Test against XML UCD version
Assuming Unicode Consortium's process for generating XML version of
the database is more reliable than my own ucd-gen, this provides us
with a way to test UCD parsing that does not depend on an up-to-date
version of ICU being available.

Note that the plaintext version of UCD is still the official source of
truth, so I'll continue using it to construct lookup tables.