Add support for NameAliases
There are no tests yet (I have to figure out how to export conflicting
GeneralCategory.Control and NameAliasType.Control first), but
experiments in repl seem to work.

The implementation is hackish in places (some integral types are fixed
to Word8 when they really should be dynamically picked), but I do not
expect many other properties to make use of the same code, and this is
already way more effort than I was willing to pour into name aliases
support.  It works for the time being, and to/fromEnum will report
errors if something doesn't fit.

All sublen arrays start with zero; this simplifies implementation at
the cost of a bit of extra memory.  Also, this way it is much easier
to orchestrate fusable list pipeline; I'm not sure I would be able to
easily do that with first offset treated specially.