Add support for canonical and compatibility decompositions
There are a couple of things in this commit worth commenting on.

For starters, the functions provided expose not the decomposition
mapping, but rather the full, recursively applied, decomposition.
This is for a couple of reasons:

1. The UAX#44 Unicode Character Database itself does not recommend
exposing raw decomposition mappings in a public interface on account
of them being too confusing without some careful reading of the docs.

2. I strongly suspect that having full decomposition will be more
useful for normalisation algorithms on strings.  They can, strictly
speaking, be written in terms of raw mappings, but the relevant part
of the code will have to utilise non-tail recursion due to the way the
mappings are set up.

The downside of that approach is that even canonical mappings no
longer have bounded length (well, in practice the bound is 4), which
necessitates the use of lists.  Surprisingly, GHC seems to be able to
optimise lists away well enough (given enough INLINE annotations),
even though I thought it didn't optimise folds of literal lists.
Apparently, I was wrong.

Second, nontrivial decompositions are exposed separately from full
decompositions.  The reasons are twofold:

1. Nontrivial decomposition required fewer branches, which might be
important in some cases.

2. I suspect that the normalisation algorithms will only call into
these functions if they already have a rough idea of what to expect,
which means that they will be able to benefit from the ‘fewer
branches’ benefit above.  But at this point this is just a conjecture
that I have no real means of verifying.

There are also no benchmarks.  First, there really isn't anything that
I can compare against (ICU does not seem to expose decompositions for
individual code points in any form).  Second, the real test here is
full normalisation algorithm; I would like to postpone optimisation
efforts until I have the full pipeline.