Replace a trie optimiser with a leaner one
First, the results: over 11x improvement in running time and almost
20% decrease in maximum residency.

The basic idea is that to evaluate the cost of a trie produced by a
specific partitioning we only really need to know two things:

1. the partitioning itself, i.e. bit counts, and
2. the number of equivalence classes at each level.

The exact formula is given by 'totalCost' function in TrieOpt, but the
gist of the idea is that each layer needs to store #{equiv. classes} *
2 ^ #{bits in between} indices of the next partitioning layer.

The algorithm itself then just precomputes these quantities for all
possible bit counts and then compares all possible partitioning
schemes rather cheaply, no longer touching any actual arrays.

Furthermore, it saves some work in equivalence class counting by
noting that representatives of equivalence classes of a coarser
partitioning contain all equivalence classes of finer partitioning.

One apparent downside of this algorithm is that we're redoing some of
deduplication work.  Unfortunately, I can't at the moment come up with
a simple way to reuse the partitions computed here in the trie
construction.  It is rather easy to do when the algorithm proceeds in
an opposite direction—from fine partitions to coarser ones—but (and
I've tried that) that tends to require auxiliary storage multiple that
of the original array, and that does not look well for our memory
budget.