Properly account for composition exclusions
While the parser handles the entirety of
DerivedNormalizationProps (well, the parts of it that aren't declared
obsolete), there is no explicit support for other properties in there
yet.

I have also expanded composition testing a bit, to catch false
negatives.

This concludes the support for Decomposition_Mapping property in this
library.  I have also struck away exclusion properties from the list,
as they seem to be strictly contributory.