Add support for joining type and group
Benchmarks show that type is a bit slower than ICU, but not too far.

I have opted to use extracted/ files, rather than original definition
in ArabicShaping.txt because they in a format that I already have a
parser for and incorporate general category-related defaults, reducing
the amount of cross-referencing I have to do.