Reduce duplication in Enum TableValue instances via DerivingVia
Of essentially Enum-based instances, Int and Bool-related are still
hand-written; that is because they are defined in GHC.Types, which is
only exposed in ghc-prim.  There's no good reason for ucd to depend on
ghc-prim (although no major harm either), so we use handwritten
instances.

Speaking of which, I could have chosen either Typeable or Generic for
reflection (and initially used Generic), but GeneralCategory and
Word32 don't have Generic instances; therefore, Typeable.