Switch all integral types to generateIntegral
This required some expansion and rethinking (e.g. now bottom layer is
not necessarily of the same type as we requested), but probably leads
to slightly cleaner results.

The benchmarks, sadly, are rather inconclusive (it seems that
criterion is not quite as good at noise removal as I'd hope), but
it seems that they're mostly changed (somewhat surprisingly, since I'd
think we'd have fewer bounds check in fromIntegral vs toEnum).