Switch to bare tagToEnum# for reading enumerations
Derived toEnum definitions are basically bounds check + tagToEnum#; we
are reading from known to be good tables, so there's no need for a
bounds check and we can safely omit it.

It's hard to judge the performance impact since everything is so
noisy, but at least in some cases where UCD was previously slower than
text-icu, it is now faster.  So I guess that's a win.

Note that this whole rigmarole really needs to be generated inline,
because tagToEnum# is deeply magical and only works if all arguments
are provided explicitly, including, most importantly, type.