Add support for Bidi_Mirroring_Glyph
In benchmarks, we slightly lose to ICU; this is *possibly* because
we're actually doing more work (the definition of `mirror` would
essentially require just omitting the conditional and the `Maybe`
wrapper) in order to stay closer to the UCD's definition of the
relevant property.  Regardless, it's not like it's the most important
property in the world (my impression is that renderers with proper
Unicode support shouldn't actually rely on it).