Add support for Bidi_Paired_Bracket(_Type)
I'm still not quite sure what is the value of Bidi_Paired_Bracket for
code points not listed in BidiBrackets.txt should be, but the xml
version of UCD uses ‘#’ for them, which means same code point
everywhere else.

Oh, and text-icu seems to have no notion of them at all.