Switch test-xml from hunit to a homegrown yocto-test framework
By sticking to the philosophy of ‘no overhead in the happy case’ and
dropping almost all features people find useful in testing
framework (but that weren't useful for me personally), yocto-test
decreases the user time by about 2x and total running time for the
test suite by 3x, from ~3 minutes to ~1.  This is still much longer
than I'd like, but it's an improvement already.