Convert a huge monolithic xml test into a group of smaller tests
Which are generally per-property now (except for numeric properties,
which are a mess), so that failure to get one property won't prevent
us from testing others.

This, by the way, is the real win of switching to Yocto: my previous
attempts at doing something broadly similar with hunit resulted with
test runs longer than anyone has any patience for.  With Yocto,
however, this whole thing runs in about the same time as monolithic
test (even slightly faster, it seems).