Delist Unicode_Radical_Stroke
Despite being listed in UAX#44, it is *very clearly* part of Unihan
and not main UCD.  Its values are not listed in non-unihan xml UCD
version, for instance.  Its values seem to be unavailable in
non-Unihan part of main UCD as well.  The actual name the property is
listed under is kRSUnicode, in line with all other Unihan properties;
the name Unicode_Radical_Stroke does not exist except as an alias and
part of UAX#44.  Maybe I'll eventually reconsider, but for now, no
thank you.