Move the package into a subdirectory
This prepares for eventual split and maybe additional packages.

In the process of trying to get dante to work with the new setup I
consolidated .dir_locals.el files, and also, unfortunately, had to
provide my own method, which is not included in this repository (it
lives in my .emacs).

Note that all executables (that means fetch.sh and ucd-gen) are
supposed to be run from the project root, whereas test suites and
benchmarks are to be run from the package directory.  This matches the
default behaviour of ‘cabal run’ and ‘cabal test/bench’.