Check in generated source files
The point is to avoid running the generator itself in every CI job,
because

1. Generation process is kind of slow, and, most importantly, consumes
boatloads of RAM (much of that is due to using boxed vectors, but
fixing that would make a number of places much more awkward).

2. I always wanted to be able to put stricter dependency bounds on the
generator than on actual library.

I didn't check the test data, though, because it's huge and boring.