Add code generator for emoji properties
The actual format of the file is similar to that of PropList, so the
Emoji module is basically a copy of that.

Emoji_Modifier is a contiguous range of 5 code points, so it will be
handled with an explicit test.

Note that the generated files won't really work as-is, as they depend
on internal UCD modules.  I'm still thinking about a best solution for
that.