Add a tentative plan for the unistring package
This is just a list of stuff that I can think of right now that I'd
like to have in the library for the first release.

Notably absent: builders.  These are complicated beasts, especially if
I want good performance (and I do, why else would anyone even use
them), so I'm not going to address them right now.