Reorganise unistring test suite
The grand idea is to have top level ‘branches’ for the kind of testing
we're doing (currently inspection and behaviour, but I want to
eventually also add allocation tests via weigh), and then mimic the
module structure of the library under each of them.  The point is to
make tests for given functionality easier to find, and keep test
modules small.