Swap allocator and storage arguments for Array
It bugged me /a lot/ that these were in different order on Allocator
class and Array proper, so now that's fixed.  It also feels,
subjectively, that the choice of representation is somehow more
fundamental than the choice of allocator, so this goes nicely with
partial application.  I think.

Anyway, Allocator class should definitely have the allocator as its
last argument because that's the way every other class works, and the
Array should keep the order for consistency.