Weaken inspection tests on toList
This highlights an unfortunate consequence of having Known storage be
a superclass of Allocator.  Namely, GHC now looks up the Known
dictionary in the Allocator one given the opportunity, instead of
resolving it from global instance and subsequently optimising away
cases.  This is sad, but in the end it seems that specialising on the
allocator is necessary for good performance anyway, and I value clean
interface more than these sort marginal gains in specialisability.