Add a small compatibility module for indexed TypeRep
It's not publicly exposed, since this sort of compatibility
shenanigans are exclusively private matter.

Speaking of which, the only reason why I even need it is because I've
decided to support GHC 8.0 and base-4.9 that comes with it.  Given
that we're nearing a release of GHC 8.10, is that /necessary/?  I
don't know, but it's here for the time being.