Give Foreign/Pinned a more sophisticated adopt method
Given that we already construct 'Foreign arrays from pinned byte
arrays, it makes perfect sense that we should be able to do that at
any later moment as well.

In fact, given that this is somewhat of a static property, it would
maybe make sense to expose this conversion as a separate infallible
primitive at some point in time.  Maybe.

Unfortunately, GHC does not have the primop we need here (unsafe thaw
of a byte array), nor does ForeignPtrContents have a constructor with
a ByteArray# payload.  So I have to resort to unsafeCoerce#, which I
believe to be safe for reasons explained in the comment.