Add convert function for Arrays
This comes with a couple of associated changes:

1. FFI for copying bytes around on GHCs that don't have appropriate
primops.

2. Every allocator monad now needs access to pointers, which means
withForeignPtr, which means they're all built on top of IO now.
Fortunately, runST and unsafeDupablePerformIO seem to have exactly
identical implementation in all supported GHC versions (via runRW#),
so we're not exactly losing much here.

Unfortunately, there are no tests for `convert` currently; this is
because I have an idea for writing them that will not require me to
hand code a billion of type combinations, but it's somewhat
complicated, and this commit is already a bundle of vaguely related
things.