Add Slice.convert
It is currently formulated as some short-cuts, followed by a catchall
conversion via Array.  The catch-alls area

1. If we can adopt the backing ByteArray#, there's no need to do
anything else (but note that we might have adopted it into
`ForeignPtr`, so we still need a generic sliceUnchecked).

2. Identity conversion on ForeignPtr's is a no-op.

The TODO worries me, but just a bit.  Essentially, the toArray
function may try to adopt the array /again/, even though we know it
will fail.  Note that there really isn't a place for function that
doesn't do this check because we don't /want/ it, so avoiding it would
probably lead to some awkward code.  Given that it /probably/ isn't
too expensive, I'm inclined to just leave it there for now.