Add a complete Scalar.Sequence.toList implementation
Using stream ‘fusion.’  I have hoped to avoid it, but it turned out to
be the most natural way to express the somewhat complicated logic in
streamLazy.  I could have probably come up with something simpler for
this particular case, but I'd like to eventually reuse this
functionality as a fallback for other functions (equal, probably
convert).