Add readWritePerformIO and readOnlyPerformIO
Together with a lengthy note explaining the reasoning behind their
implementation.  The net positive effect is that we can get rid of
custom (and ugly) unsafeDupablePerformIO' in Scalar.Sequence.

Note that readWritePerformIO is secretly identical to
unsafeDupablePerformIO, and readOnlyPerformIO is known as
inlinePerformIO in ‘text’ and accursedUnutterablePerformIO in
‘bytestring,’ but I've decided to give them names that reflect my
understanding of when they're safe to use in unistring context.